+++ tests/test_string.zig
@@ -1,9 +1,9 @@
-//! test stringIO from a user's perspective (no internal functionality)
+//! test datetime <--> string
 
 const std = @import("std");
 const builtin = @import("builtin");
 const testing = std.testing;
-const log = std.log.scoped(.zdt_test_stringIO);
+const log = std.log.scoped(.zdt_test_string);
 
 const c_locale = @cImport(@cInclude("locale.h"));
 const time_mask = switch (builtin.os.tag) {
@@ -63,12 +63,6 @@ test "format naive datetimes with format string api" {
         try testing.expectEqualStrings(case.string, buf.items);
         buf.clearAndFree();
         try case.dt.toString("%Y-%m-%d %H:%M:%S", buf.writer());
-        try testing.expectEqualStrings(case.string, buf.items);
-        buf.clearAndFree();
-        try case.dt.toString(Formats.DateOnly ++ " " ++ Formats.TimeOnly, buf.writer());
-        try testing.expectEqualStrings(case.string, buf.items);
-        buf.clearAndFree();
-        try case.dt.toString(Formats.DateTime, buf.writer());
         try testing.expectEqualStrings(case.string, buf.items);
     }
 }
@@ -516,6 +510,28 @@ test "format with %s to get Unix time in seconds" {
     }
 }
 
+test "format isocalendar, %t" {
+    const cases = [_]TestCase{
+        .{
+            .dt = try Datetime.fromFields(.{ .year = 1970, .month = 1, .day = 1 }),
+            .string = "1970-W01-4",
+            .directive = "%t",
+        },
+        .{
+            .dt = try Datetime.fromFields(.{ .year = 2024, .month = 9, .day = 21 }),
+            .string = "2024-W38-6",
+            .directive = "%t",
+        },
+    };
+
+    inline for (cases) |case| {
+        var buf = std.ArrayList(u8).init(testing.allocator);
+        try case.dt.toString(case.directive, buf.writer());
+        try testing.expectEqualStrings(case.string, buf.items);
+        buf.deinit();
+    }
+}
+
 // ---- String to Datetime ----
 
 test "comptime parse with comptime format string #1" {
@@ -836,6 +852,26 @@ test "string -> datetime -> string roundtrip with offset TZ" {
     try testing.expectEqualStrings("", std.mem.sliceTo(dt.tzinfo.?.tzOffset.?.__abbrev_data[0..], 0));
 }
 
+test "parse isocalendar, %t" {
+    const cases = [_]TestCase{
+        .{
+            .dt = try Datetime.fromFields(.{ .year = 1970, .month = 1, .day = 1 }),
+            .string = "- 1970-W01-4 -",
+            .directive = "- %t -",
+        },
+        .{
+            .dt = try Datetime.fromFields(.{ .year = 2024, .month = 9, .day = 21 }),
+            .string = "% 2024-W38-6 %",
+            .directive = "%% %t %%",
+        },
+    };
+
+    inline for (cases) |case| {
+        const dt = try Datetime.fromString(case.string, case.directive);
+        try testing.expectEqual(case.dt, dt);
+    }
+}
+
 test "parse ISO" {
     const tzutc = Tz.UTC;
     var dt_ref = try Datetime.fromFields(.{ .year = 2014, .month = 8 });
@@ -919,7 +955,7 @@ test "parse ISO" {
     try testing.expect(std.meta.eql(dt_ref, dt));
 }
 
-test "comptime parse ISO with %T" {
+test "parse ISO with %T" {
     const cases = [_]TestCase{
         .{
             .string = "2021-02-18T17:00:00.1",